+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for named-text
 
+## 1.1.5.0 -- 2024-05-?
+
+* Add HTML NameStyle and conversions between HTML and UTF8 name styles.
+* Add conversion from UTF8 to CaseInsensitive name styles (but not the reverse).
+
 ## 1.1.4.0 -- 2023-10-01
 
 * Update to allow Sayable 1.2.0.0